start33

+++ src/pages/api/clientes/clientes.js
@@ -1,27 +1,68 @@
 import connection from "@/libs/db"
+import authMiddleware from "@/middleware/authMiddleware"
 
-export default async function handler(req, res) {
-    const { usuario_id, search } = req.query;
+export async function handler(req, res) {
+
+    const user = req.user
+    if (!user) return
+
+    const { id, negocio_id, search } = req.query;
+    const isAdmin = user?.nivel === 'admin'
+    const negocioSolicitado = parseInt(negocio_id)
+    const negocioId = user?.negocio_id;
 
     if (req.method === 'GET') {
-        if (usuario_id) {
-            // Obtener un cliente por ID
+
+        if (id) {
+
             try {
                 const [rows] = await connection.query(`
                     SELECT 
-                        clientes.id, 
-                        clientes.usuario_id, 
-                        usuarios.nombre AS usuario_nombre,
-                        clientes.folio, 
-                        clientes.cliente, 
-                        clientes.contacto, 
-                        clientes.cel, 
-                        clientes.direccion, 
-                        clientes.email 
-                    FROM clientes
-                    JOIN usuarios ON clientes.usuario_id = usuarios.id 
-                    WHERE usuario_id = ?
-                    ORDER BY clientes.updatedAt DESC`, [usuario_id])
+                        c.id, 
+                        c.usuario_id, 
+                        c.usuario_nombre, 
+                        c.folio, 
+                        c.cliente, 
+                        c.contacto, 
+                        c.cel, 
+                        c.direccion, 
+                        c.email,
+                        c.negocio_id, 
+                        c.negocio_nombre
+                    FROM clientes c
+                    LEFT JOIN negocios n ON c.negocio_id = n.id
+                    ORDER BY c.updatedAt DESC`);
+                res.status(200).json(rows[0])
+            } catch (error) {
+                res.status(500).json({ error: error.message })
+            }
+
+        }
+
+        if (negocio_id) {
+
+            if (!isAdmin && negocioSolicitado !== negocioId) {
+                return res.status(403).json({ error: 'No tienes permiso para accesar' });
+            }
+
+            try {
+                const [rows] = await connection.query(`
+                    SELECT 
+                        c.id, 
+                        c.usuario_id, 
+                        c.usuario_nombre, 
+                        c.folio, 
+                        c.cliente, 
+                        c.contacto, 
+                        c.cel, 
+                        c.direccion, 
+                        c.email,
+                        c.negocio_id, 
+                        c.negocio_nombre
+                    FROM clientes c
+                    LEFT JOIN negocios n ON c.negocio_id = n.id
+                    WHERE c.negocio_id = ?
+                    ORDER BY c.updatedAt DESC`, [negocio_id])
 
                 if (rows.length === 0) {
                     /* return res.status(404).json({ error: 'Cliente no encontrado' }); */
@@ -35,78 +76,95 @@ export default async function handler(req, res) {
             return
         }
 
-            if (search) {
-                const searchQuery = `%${search.toLowerCase()}%`;
-                try {
-                    const [rows] = await connection.query(`
-                        SELECT
-                            clientes.id, 
-                            clientes.folio, 
-                            clientes.usuario_id,
-                            usuarios.nombre AS usuario_nombre,
-                            clientes.cliente,
-                            clientes.contacto,
-                            clientes.cel,
-                            clientes.direccion,
-                            clientes.email,
-                            clientes.createdAt
-                        FROM clientes
-                        JOIN usuarios ON clientes.usuario_id = usuarios.id 
-                        WHERE 
-                            LOWER(clientes.folio) LIKE ? 
-                        OR 
-                            LOWER(clientes.cliente) LIKE ?
-                        OR 
-                            LOWER(clientes.contacto) LIKE ?
-                        OR 
-                            LOWER(clientes.cel) LIKE ?
-                        OR 
-                            LOWER(clientes.direccion) LIKE ?
-                        OR 
-                            LOWER(clientes.email) LIKE ?  
-                        OR 
-                            LOWER(usuarios.nombre) LIKE ? 
-                        OR 
-                            LOWER(clientes.createdAt) LIKE ?
-                        ORDER BY clientes.updatedAt DESC`, [searchQuery, searchQuery, searchQuery, searchQuery, searchQuery, searchQuery, searchQuery, searchQuery]);
-    
-                    res.status(200).json(rows); 
-    
-                } catch (error) {
-                    res.status(500).json({ error: 'Error al realizar la búsqueda' });
+        if (search) {
+            const searchQuery = `%${search.toLowerCase()}%`;
+            try {
+
+                const whereClauses = [
+                    "LOWER(folio) LIKE ?",
+                    "LOWER(usuario_nombre) LIKE ?",
+                    "LOWER(cliente) LIKE ?",
+                    "LOWER(contacto) LIKE ?",
+                    "LOWER(direccion) LIKE ?",
+                    "LOWER(negocio_nombre) LIKE ?"
+                ];
+
+                const params = [
+                    searchQuery, searchQuery, searchQuery, searchQuery,
+                    searchQuery, searchQuery
+                ]
+
+                let whereClause = `(${whereClauses.join(" OR ")})`;
+
+                if (!isAdmin && negocioId) {
+                    whereClause += ` AND clientes.negocio_id = ?`;
+                    params.push(negocioId);
                 }
-                return;
+
+                const query = `
+                SELECT
+                  id, 
+                  usuario_id, 
+                  usuario_nombre, 
+                  folio, 
+                  cliente,
+                  contacto, 
+                  cel, 
+                  email,
+                  direccion,
+                  negocio_id,
+                  negocio_nombre,
+                  createdAt
+                FROM clientes
+                WHERE ${whereClause}
+                ORDER BY updatedAt DESC
+              `;
+
+                const [rows] = await connection.query(query, params);
+                return res.status(200).json(rows);
+
+            } catch (error) {
+                console.error(error)
+                return res.status(500).json({ error: 'Error al realizar la búsqueda' });
             }
 
-            // Obtener todos los clientes
+        } else {
+
+            if (!isAdmin) {
+                return res.status(403).json({ error: 'No tienes permiso para accesar' });
+            }
+
             try {
                 const [rows] = await connection.query(`
                     SELECT 
-                        clientes.id, 
-                        clientes.usuario_id, 
-                        usuarios.nombre AS usuario_nombre,
-                        clientes.folio, 
-                        clientes.cliente, 
-                        clientes.contacto, 
-                        clientes.cel, 
-                        clientes.direccion, 
-                        clientes.email 
-                    FROM clientes
-                    JOIN usuarios ON clientes.usuario_id = usuarios.id 
-                    ORDER BY clientes.updatedAt DESC`);
+                        c.id, 
+                        c.usuario_id, 
+                        c.usuario_nombre, 
+                        c.folio, 
+                        c.cliente, 
+                        c.contacto, 
+                        c.cel, 
+                        c.direccion, 
+                        c.email,
+                        c.negocio_id, 
+                        c.negocio_nombre
+                    FROM clientes c
+                    LEFT JOIN negocios n ON c.negocio_id = n.id
+                    ORDER BY c.updatedAt DESC`);
                 res.status(200).json(rows)
             } catch (error) {
                 res.status(500).json({ error: error.message })
             }
-        
+        }
     } else if (req.method === 'POST') {
+
         try {
-            const { usuario_id, folio, cliente, contacto, cel, direccion, email } = req.body
-            if (!cliente ) {
+            const { usuario_id, usuario_nombre, folio, cliente, contacto, cel, direccion, email, negocio_id, negocio_nombre } = req.body
+            if (!cliente) {
                 return res.status(400).json({ error: 'Todos los datos son obligatorios' })
             }
 
-            const [result] = await connection.query('INSERT INTO clientes (usuario_id, folio, cliente, contacto, cel, direccion, email) VALUES (?, ?, ?, ?, ?, ?, ?)', [usuario_id, folio, cliente, contacto, cel, direccion, email])
+            const [result] = await connection.query('INSERT INTO clientes (usuario_id, usuario_nombre, folio, cliente, contacto, cel, direccion, email, negocio_id, negocio_nombre) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?)', [usuario_id, usuario_nombre, folio, cliente, contacto, cel, direccion, email, negocio_id, negocio_nombre])
             const newClient = { id: result.insertId }
             res.status(201).json(newClient)
         } catch (error) {
@@ -114,55 +172,82 @@ export default async function handler(req, res) {
         }
     } else if (req.method === 'PUT') {
 
+        /* if (!isAdmin && negocioSeleccionado !== negocioId) {
+            return res.status(403).json({ error: 'No tienes permiso para modificar este recurso' })
+          } */
+
         const { id } = req.query;
 
         if (!id) {
             return res.status(400).json({ error: 'ID del cliente es obligatorio' })
         }
 
-        const { folio, cliente, contacto, cel, direccion, email } = req.body
+        const { cliente, contacto, cel, direccion, email } = req.body
 
         if (!cliente) {
             return res.status(400).json({ error: 'Todos los datos son obligatorios' })
         }
 
         try {
-            const [result] = await connection.query('UPDATE clientes SET folio = ?, cliente = ?, contacto = ?, cel = ?, direccion = ?, email = ? WHERE id = ?', [folio, cliente, contacto, cel, direccion, email, id])
+            const [result] = await connection.query('UPDATE clientes SET cliente = ?, contacto = ?, cel = ?, direccion = ?, email = ? WHERE id = ?', [cliente, contacto, cel, direccion, email, id])
 
             if (result.affectedRows === 0) {
                 return res.status(404).json({ error: 'Cliente no encontrado' })
             }
 
-            res.status(200).json({ message: 'Cliente actualizado correctamente' })
+            // También actualizamos notas que tienen este cliente_id
+            await connection.query(`
+            UPDATE notas 
+            SET cliente_nombre = ?, cliente_contacto = ?
+            WHERE cliente_id = ?
+            `, [cliente, contacto, id])
+
+            res.status(200).json({ message: 'Cliente y notas actualizados correctamente' })
+
         } catch (error) {
             res.status(500).json({ error: error.message })
         }
     } else if (req.method === 'DELETE') {
-
         const { id } = req.query;
 
-
         if (!id) {
-            return res.status(400).json({ error: 'ID del cliente es obligatorio' })
+            return res.status(400).json({ error: 'ID del cliente es obligatorio' });
         }
 
         try {
-            // Actualizar la columna client_id en la tabla notas a NULL para el cliente que se va a eliminar
-            await connection.query('UPDATE notas SET cliente_id = NULL WHERE cliente_id = ?', [id])
+            // Primero obtenemos el negocio_id del cliente que se quiere eliminar
+            const [[cliente]] = await connection.query(
+                'SELECT negocio_id FROM clientes WHERE id = ?',
+                [id]
+            );
 
-            // Ahora eliminar el cliente
-            const [result] = await connection.query('DELETE FROM clientes WHERE id = ?', [id])
-
-            if (result.affectedRows === 0) {
-                return res.status(404).json({ error: 'Cliente no encontrado' })
+            if (!cliente) {
+                return res.status(404).json({ error: 'Cliente no encontrado' });
             }
 
-            res.status(200).json({ message: 'Cliente eliminado correctamente' })
+            const negocioSeleccionado = cliente.negocio_id;
+
+            // Validación de permisos
+            if (!isAdmin && negocioSeleccionado !== negocioId) {
+                return res.status(403).json({ error: 'No tienes permiso para modificar este recurso' });
+            }
+
+            // Actualizar notas que usan a este cliente
+            await connection.query('UPDATE notas SET cliente_id = NULL WHERE cliente_id = ?', [id]);
+
+            // Eliminar cliente
+            const [result] = await connection.query('DELETE FROM clientes WHERE id = ?', [id]);
+
+            if (result.affectedRows === 0) {
+                return res.status(404).json({ error: 'Cliente no encontrado' });
+            }
+
+            res.status(200).json({ message: 'Cliente eliminado correctamente' });
         } catch (error) {
-            res.status(500).json({ error: error.message })
+            res.status(500).json({ error: error.message });
         }
-    } else {
-        res.setHeader('Allow', ['GET', 'POST', 'PUT', 'DELETE'])
-        res.status(405).end(`Method ${req.method} Not Allowed`)
     }
+
 }
+
+export default authMiddleware(handler)
